@@ -1,16 +1,16 @@
 {
     "name": "hanaboso/data-grid",
     "description": "Data grid",
-    "version": "2.1.9",
+    "version": "2.2.0",
     "license": "Apache-2.0",
     "require": {
-        "php": "^7.4",
+        "php": "^8.0",
         "ext-ctype": "*",
         "ext-intl": "*",
         "ext-mbstring": "*",
-        "beberlei/doctrineextensions": "^1.2",
-        "doctrine/orm": "^2.7",
-        "hanaboso/utils": "^1.0"
+        "beberlei/doctrineextensions": "^1.3",
+        "doctrine/orm": "^2.8",
+        "hanaboso/utils": "^1.1"
     },
     "require-dev": {
         "hanaboso/php-check-utils": "^1.1"
